move account provider APIs to client crate for RPC
