+++ client/Cargo.toml
@@ -16,6 +16,10 @@ parking_lot = "0.8"
 panic_hook = { git = "https://github.com/Conflux-Chain/conflux-parity-deps.git" }
 io = { path = "../util/io" }
 ethkey = { path = "../accounts/ethkey" }
+ethstore = { path = "../accounts/ethstore" }
+ethcore-accounts = { path = "../accounts" }
+app_dirs = { git = "https://github.com/paritytech/app-dirs-rs" }
+dir = { git = "https://github.com/Conflux-Chain/conflux-parity-deps.git" }
 ctrlc = { git = "https://github.com/paritytech/rust-ctrlc.git" }
 jsonrpc-core = "12.0.0"
 jsonrpc-tcp-server = "=12.0.0"
